Embedding/requirements.txt
supabase==2.3.4
pinecone-client==3.0.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
trafilatura==8.0.1
numpy==1.24.3 